Concepts

Description

This module defines data types and algorithms for abstractly manipulating concepts. The exported functions are meant to be used as

>>> putStrLn $ showConcepts $ simplify $ singleton $ ForwardIterator (Type "I")
Valid ValueType<I>
EqualityComparable I
Semiregular I
Regular I
Readable I
WeaklyIncrementable I
Incrementable I

Documentation

data Concept

This data type represents an abstract requirement.

Constructors

Valid TypeExpr
This concept is the most basic requirement: it's argument must be a valid type.

Same TypeExpr TypeExpr
Same<T, U>

Derived TypeExpr TypeExpr
Derived<T, U>

Convertible TypeExpr TypeExpr
Convertible<T, U>

Common TypeExpr TypeExpr
Common<T, U>

Integral TypeExpr
Integral<T>

SignedIntegral TypeExpr
SignedIntegral<T>

UnsignedIntegral TypeExpr
UnsignedIntegral<T>

EqualityComparable TypeExpr
EqualityComparable<T>

EqualityComparable2 TypeExpr TypeExpr
EqualityComparable<T, U>

WeaklyOrdered TypeExpr
WeaklyOrdered<T>

WeaklyOrdered2 TypeExpr TypeExpr
WeaklyOrdered<T, U>

TotallyOrdered TypeExpr
TotallyOrdered<T>

TotallyOrdered2 TypeExpr TypeExpr
TotallyOrdered<T, U>

Semiregular TypeExpr
Semiregular<T>

Regular TypeExpr
Regular<T>

Function TypeExpr [TypeExpr]
Function<F, Args...>

RegularFunction TypeExpr [TypeExpr]
RegularFunction<F, Args...>

Predicate TypeExpr [TypeExpr]
Predicate<P, Args...>

Relation TypeExpr TypeExpr
Relation<R, T>

Relation2 TypeExpr TypeExpr TypeExpr
Relation<R, T, U>

UnaryOperation TypeExpr TypeExpr
UnaryOperation<Op, T>

BinaryOperation TypeExpr TypeExpr
BinaryOperation<Op, T>

BinaryOperation2 TypeExpr TypeExpr TypeExpr
BinaryOperation<Op, T, U>

Readable TypeExpr

MoveWritable TypeExpr TypeExpr

Writable TypeExpr TypeExpr

IndirectlyMovable TypeExpr TypeExpr

IndirectlyCopyable TypeExpr TypeExpr

WeaklyIncrementable TypeExpr

Incrementable TypeExpr

Decrementable TypeExpr

Advancable TypeExpr

WeakInputIterator TypeExpr

InputIterator TypeExpr

ForwardIterator TypeExpr

BidirectionalIterator TypeExpr

RandomAccessIterator TypeExpr

Permutable TypeExpr

Mergeable TypeExpr TypeExpr TypeExpr

Mergeable2 TypeExpr TypeExpr TypeExpr TypeExpr

Sortable TypeExpr

Sortable2 TypeExpr TypeExpr

data TypeExpr

Constructors

Type String

commonType :: TypeExpr -> TypeExpr -> TypeExpr

Performs the type function CommonType<T, U>

valueType :: TypeExpr -> TypeExpr

Performs the type function ValueType<T>

resultOf :: TypeExpr -> [TypeExpr] -> TypeExpr

Performs the type function ResultOf<F, Args...>

satisfies :: Set Concept -> Set Concept -> Bool

Checks if the first concept set satisfies the requirements of the second. Will fully simplify both arguments. Ex.

>>> let i = Type "I"
>>> satisfies (singleton $ ForwardIterator i) (singleton $ InputIterator i)
True

showConcepts :: Set Concept -> String

A prettier display function for showing sets of concepts

simplify :: Set Concept -> Set Concept

Performs a fixed-point reduction on the set of concepts

simplify' :: Concept -> [Concept]

Performs one reduction step upon a concept. This is the underlying operation used by simplify.